修改代理方法，能够自主选择sql语句的执行类型

+++ src/test/java/cn/bobasyu/mybatis/test/mapper/IUserMapper.java
@@ -3,5 +3,5 @@ package cn.bobasyu.mybatis.test.mapper;
 import cn.bobasyu.mybatis.test.entity.User;
 
 public interface IUserMapper {
-    User queryUserInfoById(Long id);
+    User queryUserInfoById(Integer id);
 }
